Real-Time Camera Face Counting with Image Processing
It is aimed to count the faces in images taken from a camera in real time using OpenCV, Haar Cascade classifier and image processing techniques. Face detections were performed on the images taken via the camera and the number of faces was displayed on the screen instantly. This process was carried out using a ready-made machine learning model and image processing techniques.